requirements.txt
functions-framework==3.2.0
google-api-python-client==2.66.0
google-cloud-storage==2.6.0